[core] De-duplicate shader prelude source
